fix: oversized repository lines
The repository would stretch to fill the entire app viewport. This is because of the flex-1 class in tailwind allowing the flex item to stretch to fill the remaining space.

Changing to flex-initial prevents the stretch. This could also be removed entirely but I thought it's best to have something set so we're not relying on electron to set some defaults.